add constant.js
