+++ src/helpers/index.js
@@ -29,4 +29,4 @@ export const randomString = (len, an) => {
 
 export const makeID = () => `ISBN-${randomString(10, 'n')}`;
 
-export const CATEGORIES = ['Action', 'Biography', 'History', 'Horror', 'Kids', 'Learning', 'Sci-Fi'];
+
